+++ src/grid.h
@@ -5,11 +5,11 @@
 #include "dimension.h"
 #include "vector_sparse.h"
 
-class Grid : public std::vector<double> {
+class Grid : public std::vector<float> {
 public:
     void set(const Dimension &latitude, const Dimension &longitude, const Dimension &time);
-    VectorSparse basis(const double x, const double y, const double z) const;
-    double operator ()(const double x, const double y, const double z) const;
+    VectorSparse basis(const float x, const float y, const float z) const;
+    float operator ()(const float x, const float y, const float z) const;
 private:
     Dimension latitude, longitude, time;
 };
