+++ src/ray.cpp
@@ -13,7 +13,7 @@ std::ostream& operator << (std::ostream &out, Ray c) {
     return out;
 }
 
-Ray::Ray(const point &station, const point &satellite, const double time) {
+Ray::Ray(const point &station, const point &satellite, const float time) {
     this->station = station;
     this->satellite = satellite;
     this->time = time;
@@ -21,17 +21,17 @@ Ray::Ray(const point &station, const point &satellite, const double time) {
 
 void Ray::compute_cross() {
     const point dr = satellite - station;
-    const double rdr = station * dr;
-    const double t = (-rdr + sqrt(rdr * rdr + (dr * dr) * ((Re + h) * (Re + h) - (station * station)))) / (dr * dr);
+    const float rdr = station * dr;
+    const float t = (-rdr + sqrt(rdr * rdr + (dr * dr) * ((Re + h) * (Re + h) - (station * station)))) / (dr * dr);
     cross = station + t * (satellite - station);
 }
 
 void Ray::compute_angles() {
     const point diff = (satellite - station);
-    const double a = cross * diff;
-    const double c = sqrt((cross * cross) * (diff * diff));
-    const double ac = fmin(a / c, 1);
-    angle = acos(ac);
+    const float a = cross * diff;
+    const float c = std::sqrt((cross * cross) * (diff * diff));
+    const float ac = std::fmin(a / c, 1.0f);
+    angle = std::acos(ac);
     thetta = asin(cross.R[2] / (Re + h));
     phi = atan2(cross.R[1], cross.R[0]);
 }
