+++ src/reconstruction.cpp
@@ -4,10 +4,10 @@
 #include "reconstruction.h"
 
 
-void Art::operator()(Grid &x, const std::vector<VectorSparse> &a, const std::vector<double> &m, bool onlyPositive) const {
+void Art::operator()(Grid &x, const std::vector<VectorSparse> &a, const std::vector<float> &m, bool onlyPositive) const {
     for (unsigned i = 0; i < m.size(); ++i) {
-        double aa = 0;
-        double ax = 0;
+        float aa = 0;
+        float ax = 0;
         for (unsigned k = 0; k < a[i].size(); ++k) {
             const auto [index, value] = a[i][k];
             aa += value * value;
@@ -16,25 +16,25 @@ void Art::operator()(Grid &x, const std::vector<VectorSparse> &a, const std::vec
         if (aa == 0) {
             continue;
         }
-        const double t = (m[i] - ax) / aa;
+        const float t = (m[i] - ax) / aa;
         for (unsigned j = 0; j < a[i].size(); ++j) {
             const auto [index, value] = a[i][j];
             x[index] += value * t;
         }
         if (onlyPositive) {
             std::transform(x.begin(), x.end(), x.begin(),
-                [] (double value) { return std::max(value, 0.0); }
+                [] (float value) { return std::max(value, 0.f); }
             );
         }
     }
 
 }
 
-void Sirt::operator()(Grid &x, const std::vector<VectorSparse> &a, const std::vector<double> &m, bool onlyPositive) const {
+void Sirt::operator()(Grid &x, const std::vector<VectorSparse> &a, const std::vector<float> &m, bool onlyPositive) const {
     // Calculating x increment
-    std::vector<double> dx(x.size(), 0);
+    std::vector<float> dx(x.size(), 0);
     for (unsigned i = 0; i < m.size(); ++i) {
-        double ax = 0;
+        float ax = 0;
         for (unsigned j = 0; j < a[i].size(); ++j) {
             const auto [index, value] = a[i][j];
             ax += value * x[index];
@@ -46,11 +46,11 @@ void Sirt::operator()(Grid &x, const std::vector<VectorSparse> &a, const std::ve
     }
 
     // Calculating t coefficient for residual minimization
-    std::vector<double> adx;
-    std::vector<double> axy;
+    std::vector<float> adx;
+    std::vector<float> axy;
     for (unsigned j = 0; j < m.size(); ++j) {
-        double adxj = 0;
-        double axyj = 0;
+        float adxj = 0;
+        float axyj = 0;
         for (unsigned k = 0; k < a[j].size(); ++k) {
             const auto [index, value] = a[j][k];
             adxj += value * dx[index];
@@ -62,16 +62,16 @@ void Sirt::operator()(Grid &x, const std::vector<VectorSparse> &a, const std::ve
 
     // Numerator is a sum of axy and adx product values
     // While denominator is a sum of adx values squared
-    const double numerator = std::inner_product(axy.begin(), axy.end(), adx.begin(), 0.0);
-    const double denominator = std::inner_product(adx.begin(), adx.end(), adx.begin(), 0.0);
-    const double t = - numerator / denominator;
+    const float numerator = std::inner_product(axy.begin(), axy.end(), adx.begin(), 0.f);
+    const float denominator = std::inner_product(adx.begin(), adx.end(), adx.begin(), 0.f);
+    const float t = - numerator / denominator;
 
     std::transform(dx.begin(), dx.end(), dx.begin(), [t](auto i) { return t * i;});
-    std::transform(x.begin(), x.end(), dx.begin(), x.begin(), std::plus<double>());
+    std::transform(x.begin(), x.end(), dx.begin(), x.begin(), std::plus<float>());
 
     if (onlyPositive) {
         std::transform(x.begin(), x.end(), x.begin(),
-            [] (double value) { return std::max(value, 0.0); }
+            [] (float value) { return std::max(value, 0.f); }
         );
     }
 }
