+++ src/reconstruction.h
@@ -8,17 +8,17 @@
 class Solver {
 public:
     Solver() {}
-    virtual void operator()(Grid &grid, const std::vector<VectorSparse> &a, const std::vector<double> &m, bool onlyPositive=true) const = 0;
+    virtual void operator()(Grid &grid, const std::vector<VectorSparse> &a, const std::vector<float> &m, bool onlyPositive=true) const = 0;
     virtual ~Solver() {}
 };
 
 
 class Art : public Solver {
 public:
-    void operator()(Grid &grid, const std::vector<VectorSparse> &a, const std::vector<double> &m, bool onlyPositive=true) const;
+    void operator()(Grid &grid, const std::vector<VectorSparse> &a, const std::vector<float> &m, bool onlyPositive=true) const;
 };
 
 class Sirt : public Solver {
 public:
-    void operator()(Grid &grid, const std::vector<VectorSparse> &a, const std::vector<double> &m, bool onlyPositive=true) const;
+    void operator()(Grid &grid, const std::vector<VectorSparse> &a, const std::vector<float> &m, bool onlyPositive=true) const;
 };
