+++ src/spline.h
@@ -4,14 +4,14 @@
 
 class Spline{
 public:
-    Spline(const std::vector<double> x, const std::vector<double> y);
+    Spline(const std::vector<float> x, const std::vector<float> y);
     ~Spline() = default;
-    double operator() (const double x) const; // return value at point x
+    float operator() (const float x) const; // return value at point x
 private:
     // structure that defines spline on each segment
     struct SplineTuple {
-        double a, b, c, d, x;
+        float a, b, c, d, x;
     };
     std::vector<SplineTuple> splines;
-    void build(const std::vector<double> x, const std::vector<double> y);
+    void build(const std::vector<float> x, const std::vector<float> y);
 };
